Refactor shared logic into common module
