+++ app_test.py
@@ -1,9 +1,8 @@
 import streamlit as st
-import streamlit.components.v1 as components
 import pandas as pd
 import unicodedata
 from io import StringIO
-import os
+from common import calculate_plan, format_euro, produtos, setup_page
 
 
 def normalize(text: str) -> str:
@@ -13,95 +12,10 @@ def normalize(text: str) -> str:
     ).lower()
 
 # Força tema claro no Streamlit
-st.set_page_config(layout="centered")
-
-# Estilo personalizado com fundo branco e tudo a preto
-st.markdown("""
-    <style>
-        @import url('https://fonts.googleapis.com/css2?family=Segoe+UI&display=swap');
-
-        html, body, [class*="css"] {
-            font-family: 'Segoe UI', sans-serif;
-            background-color: #FFFFFF !important;
-        }
-
-        .stApp {
-            background-color: #FFFFFF !important;
-        }
-
-        body, h1, h2, h3, h4, p, label, span,
-        .stCheckbox>div, .stCheckbox span, .stCheckbox label span,
-        .stSelectbox label, .stNumberInput label,
-        .stMarkdown span, .stMarkdown p, .stMarkdown div,
-        .css-1x8cf1d, .css-1v0mbdj span, .css-17eq0hr, .css-1r6slb0 {
-            color: #000000 !important;
-            font-weight: 500 !important;
-            opacity: 1 !important;
-@@ -73,217 +81,466 @@ st.markdown("""
-
-        .stSuccess {
-            background-color: #c9cfd3 !important;
-            border-left: 6px solid #0046FE !important;
-            color: #000000 !important;
-        }
-
-        .logo-container {
-            display: flex;
-            justify-content: center;
-            margin-bottom: 20px;
-        }
-    </style>
-""", unsafe_allow_html=True)
-
-st.markdown("""
-    <div class="logo-container">
-        <img src="https://phcsoftware.com/pt/wp-content/uploads/sites/3/2023/11/logo.svg" width="220" />
-    </div>
-""", unsafe_allow_html=True)
+setup_page()
 
 st.title("Simulador de Plano PHC Evolution")
 
-# Produtos com planos mínimos
-produtos = {
-    "Core e Transversais": {
-        "Inventário Avançado": {"plano": 3, "per_user": False},
-        "Frota": {"plano": 3, "per_user": False},
-        "Logística": {"plano": 5, "per_user": False},
-        "Denúncias": {"plano": 5, "per_user": False},
-        "Documentos": {"plano": 3, "per_user": False},
-        "GenAI": {"plano": 2, "per_user": False},
-        "CRM": {"plano": 3, "per_user": True},
-        "BPM": {"plano": 5, "per_user": False},
-        "Ponto de Venda (POS/Restauração)": {"plano": 1, "per_user": True},
-    },
-    "Área Financeira e Recursos Humanos": {
-        "Contabilidade": {"plano": 3, "per_user": True},
-        "Ativos": {"plano": 3, "per_user": True},
-        "Vencimento": {"plano": 3, "per_user": True},
-        "Colaborador": {"plano": 5, "per_user": True},
-        "Careers c/ Recrutamento": {"plano": 5, "per_user": True},
-        "OKR": {"plano": 4, "per_user": True},
-        "Formação": {"plano": 3, "per_user": False},
-        "Imóveis": {"plano": 3, "per_user": False},
-    },
-    "Outros": {
-        "Suporte": {"plano": 2, "per_user": True},
-        "Ecommerce B2B": {"plano": 3, "per_user": False},
-    },
-    "Projeto": {
-        "Orçamentação": {"plano": 3, "per_user": True},
-        "Orçamentação + Medição": {"plano": 3, "per_user": True},
-        "Orçamentação + Medição + Controlo": {"plano": 3, "per_user": True},
-        "Full Project - Controlo + Medição + Orçamentação + Planeamento + Revisão de Preços": {
-            "plano": 3,
-            "per_user": True,
-        },
-    },
-    "Connected Services": {
-        "Bank Connector": {"plano": 4, "per_user": False},
-        "EDI Broker": {"plano": 1, "per_user": False},
-    },
-}
 
 # Área para colar tabela do Excel (opcional)
 with st.expander("Importar tabela", expanded=False):
@@ -288,7 +202,6 @@ utilizadores = st.number_input(
 
 # Captura das seleções
 selecoes = {}
-bank_connector_selecionado = False
 for area, modulos in produtos.items():
     with st.expander(area, expanded=False):
         if area == "Projeto":
@@ -318,7 +231,6 @@ for area, modulos in produtos.items():
                     st.markdown(
                         "O Plano Advanced inclui ligação a 1 Banco, o Plano Premium a 3 Bancos e o Ultimate a 5 Bancos, se precisar de mais bancos além dos incluídos, indique o nº necessário"
                     )
-                    bank_connector_selecionado = ativado
                     if ativado:
                         selecoes[modulo] = st.number_input(
                             "Nº Bancos Adicionais",
@@ -342,134 +254,35 @@ for area, modulos in produtos.items():
 
 # Lógica do plano
 if st.button("Calcular Plano Recomendado"):
-    planos = []
+    resultado = calculate_plan(
+        plano_atual,
+        tipo_gestao,
+        utilizadores,
+        selecoes,
+        extras_importados,
+        extras_planos,
+    )
 
-    if plano_atual == "Enterprise":
-        planos.append(6)
-    elif plano_atual == "Advanced":
-        planos.append(4)
-    elif plano_atual == "Corporate":
-        planos.append(1)
-        if tipo_gestao == "Gestão Terceiros":
-            planos.append(2)
-        elif tipo_gestao == "Gestão Completo":
-            planos.append(3)
+    for msg in resultado["warnings"]:
+        st.warning(msg)
 
-    csv_path = "precos_planos.csv"
-    df_precos = pd.read_csv(csv_path, sep=",")
-    df_produtos = pd.read_csv("precos_produtos.csv", sep=",")
-
-    limites = [
-        (int(row["plano_id"]), row.get("limite_utilizadores"))
-        for _, row in df_precos.iterrows()
-    ]
-    limites.sort(key=lambda x: x[0])
-
-    plano_utilizadores = None
-    for pid, limite in limites:
-        if pd.notna(limite) and str(limite).strip() != "":
-            if utilizadores <= int(limite):
-                plano_utilizadores = pid
-                break
-    if plano_utilizadores is None:
-        plano_utilizadores = max(pid for pid, _ in limites)
-
-    planos.append(plano_utilizadores)
-
-    for modulo in selecoes:
-        plano_min = None
-        for area in produtos.values():
-            if modulo in area:
-                info = area[modulo]
-                plano_min = info.get("plano")
-                break
-        if plano_min:
-            planos.append(plano_min)
-    for extra_mod in extras_importados:
-        planos.append(extras_planos.get(extra_mod, 0))
-    if "Colaborador" in selecoes and "Vencimento" not in selecoes:
-        st.warning("O módulo Colaborador requer Vencimento")
-
-    plano_final = max(planos) if planos else 1
-    preco_planos = {
-        int(row["plano_id"]): (
-            row["nome"],
-            float(row.get("preco_base", 0) or 0),
-            int(row.get("utilizadores_incluidos", 0) or 0),
-            float(row.get("preco_extra_ate_10", 0)) if str(row.get("preco_extra_ate_10", "")).strip() else 0,
-            float(row.get("preco_extra_ate_50", 0)) if str(row.get("preco_extra_ate_50", "")).strip() else 0,
-            float(row.get("preco_extra_acima_50", 0)) if str(row.get("preco_extra_acima_50", "")).strip() else 0
-        )
-        for _, row in df_precos.iterrows()
-    }
-
-    nome, preco_base, incluidos, preco_ate_10, preco_ate_50, preco_mais_50 = preco_planos[plano_final]
-
-    preco_produtos = {
-        (row["produto"], int(row["plano_id"])): (
-            float(row.get("preco_base", 0) or 0),
-            float(row.get("preco_unidade", 0) or 0),
-        )
-        for _, row in df_produtos.iterrows()
-    }
-
-    custo_extra_utilizadores = 0
-    extras = max(0, utilizadores - incluidos)
-    grupo1 = grupo2 = grupo3 = 0
-
-    if extras > 0:
-        if plano_final == 6:
-            grupo1 = min(5, extras)
-            grupo2 = min(40, max(0, extras - 5))
-            grupo3 = max(0, extras - 45)
-            custo_extra_utilizadores = grupo1 * preco_ate_10 + grupo2 * preco_ate_50 + grupo3 * preco_mais_50
-        else:
-            grupo1 = extras
-            custo_extra_utilizadores = grupo1 * preco_ate_10
-
-    custo_modulos = 0
-    modulos_detalhe = {}
-    for modulo, quantidade in selecoes.items():
-        if modulo == "Ponto de Venda (POS/Restauração)":
-            preco_primeiro = preco_produtos.get(("POS (1º)", plano_final), (0, 0))[0]
-            preco_2_10 = preco_produtos.get(("POS (2 a 10)", plano_final), (0, 0))[1]
-            preco_maior_10 = preco_produtos.get(("POS (>10)", plano_final), (0, 0))[1]
-
-            if quantidade > 0:
-                restantes = quantidade - 1
-                ate_10 = min(restantes, 9)
-                acima_10 = max(restantes - 9, 0)
-                custo_base = preco_primeiro
-                custo_extra = ate_10 * preco_2_10 + acima_10 * preco_maior_10
-            else:
-                custo_base = 0
-                custo_extra = 0
-        else:
-            base, unidade = preco_produtos.get((modulo, plano_final), (0, 0))
-            custo_base = base
-            custo_extra = unidade * quantidade if unidade else 0
-
-        if custo_base or custo_extra:
-            custo_total = custo_base + custo_extra
-            custo_modulos += custo_total
-            modulos_detalhe[modulo] = (custo_base, custo_extra)
-
-    custo_estimado = preco_base + custo_extra_utilizadores + custo_modulos
-
-    st.success(f"Plano PHC Evolution recomendado: {nome}")
-    def format_euro(valor: float) -> str:
-        return f"{int(round(valor)):,}".replace(",", ".") + " €"
-
-    st.markdown(f"**Previsão de Custo do Plano:** {format_euro(custo_estimado)}")
+    st.success(f"Plano PHC Evolution recomendado: {resultado['nome']}")
+    st.markdown(
+        f"**Previsão de Custo do Plano:** {format_euro(resultado['custo_estimado'])}"
+    )
 
     detalhes = []
-    detalhes.append(("Preço do Plano Base", format_euro(preco_base), False))
-    if custo_extra_utilizadores > 0:
+    detalhes.append(("Preço do Plano Base", format_euro(resultado["preco_base"]), False))
+    if resultado["custo_extra_utilizadores"] > 0:
         detalhes.append(
-            (f"Preço dos {extras} Full Users adicionais", format_euro(custo_extra_utilizadores), True)
+            (
+                f"Preço dos {resultado['extras_utilizadores']} Full Users adicionais",
+                format_euro(resultado["custo_extra_utilizadores"]),
+                True,
+            )
         )
 
-    for modulo, custos in modulos_detalhe.items():
+    for modulo, custos in resultado["modulos_detalhe"].items():
         custo_base, custo_extra = custos
         detalhes.append((modulo, format_euro(custo_base), False))
         if custo_extra > 0:
@@ -485,19 +298,11 @@ if st.button("Calcular Plano Recomendado"):
             unsafe_allow_html=True,
         )
 
-    if bank_connector_selecionado:
-        bancos_base = 0
-        if plano_final == 4:
-            bancos_base = 1
-        elif plano_final == 5:
-            bancos_base = 3
-        elif plano_final == 6:
-            bancos_base = 5
-        if bancos_base:
-            st.markdown(
-                f"<p style='color:#000000;'>Bank Connector inclui {bancos_base} banco(s) base.</p>",
-                unsafe_allow_html=True,
-            )
+    if resultado["bancos_base"]:
+        st.markdown(
+            f"<p style='color:#000000;'>Bank Connector inclui {resultado['bancos_base']} banco(s) base.</p>",
+            unsafe_allow_html=True,
+        )
 
     if "genai" in extras_importados:
         st.markdown(
